requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gptcache==0.1.11
idna==3.4
importlib-metadata==6.4.1
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.141
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==1.5.3
pdfminer.six==20201018
pdfx==1.4.1
Pillow==9.5.0
pinecone-client==2.2.1
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==13.3.4
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.47
streamlit==1.21.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
yarl==1.8.2
zipp==3.15.0
tiktoken==0.3.3
